requirements.txt
fastapi
uvicorn
chromadb 
sentence-transformers 
pypdf 
PyPDF2
unstructured 
openai 
langchain
streamlit 
requests
pymupdf
markdown
python-docx
llama-cpp-python
huggingface-hub
transformers 
accelerate 
bitsandbytes 
sentencepiece